requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
altair==4.2.0
    # via -r requirements/requirements.in
attrs==22.1.0
    # via jsonschema
black==22.6.0
    # via -r requirements/requirements.in
click==8.1.3
    # via
    #   -r requirements/requirements.in
    #   black
colorama==0.4.5
    # via gamma-config
commonmark==0.9.1
    # via rich
cycler==0.11.0
    # via matplotlib
entrypoints==0.4
    # via altair
et-xmlfile==1.1.0
    # via openpyxl
fonttools==4.34.4
    # via matplotlib
fsspec==2022.5.0
    # via -r requirements/requirements.in
gamma-config==0.5.0
    # via -r requirements/requirements.in
gurobipy==9.1.2
    # via -r requirements/requirements.in
jinja2==3.1.2
    # via altair
joblib==1.1.0
    # via scikit-learn
jsonschema==4.9.1
    # via altair
kiwisolver==1.4.4
    # via matplotlib
markupsafe==2.1.1
    # via jinja2
matplotlib==3.5.2
    # via seaborn
mypy-extensions==0.4.3
    # via black
numpy==1.23.1
    # via
    #   altair
    #   matplotlib
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
    #   seaborn
    #   xgboost
openpyxl==3.0.10
    # via -r requirements/requirements.in
packaging==21.3
    # via matplotlib
pandas==1.4.3
    # via
    #   -r requirements/requirements.in
    #   altair
    #   seaborn
pathspec==0.9.0
    # via black
pillow==9.2.0
    # via matplotlib
platformdirs==2.5.2
    # via black
pyarrow==8.0.0
    # via -r requirements/requirements.in
pydantic==1.9.1
    # via -r requirements/requirements.in
pygments==2.12.0
    # via rich
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   gamma-config
    #   matplotlib
    #   pandas
python-dotenv==0.20.0
    # via gamma-config
pytz==2022.1
    # via pandas
rich==12.4.4
    # via -r requirements/requirements.in
ruamel-yaml==0.17.21
    # via gamma-config
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
scikit-learn==1.1.2
    # via sklearn
scipy==1.9.0
    # via
    #   scikit-learn
    #   seaborn
    #   xgboost
seaborn==0.11.2
    # via -r requirements/requirements.in
six==1.16.0
    # via python-dateutil
sklearn==0.0
    # via -r requirements/requirements.in
threadpoolctl==3.1.0
    # via scikit-learn
tomli==2.0.1
    # via black
toolz==0.12.0
    # via altair
typing-extensions==4.3.0
    # via
    #   black
    #   pydantic
xgboost==1.6.1
    # via -r requirements/requirements.in
xlrd==2.0.1
    # via -r requirements/requirements.in
